+++ tests/unit_tests/test_connections.py
@@ -81,10 +81,8 @@ class TestConst(TestCase):
         self.assertFalse(res.is_error())
 
     def test_masked_opts(self):
-        res = mask_opts({"headers": {"Authorization": "Bearer 123456"}, "data": b'0' * 350, "method": "GET"})
-        self.assertEqual(res,
-                         {"headers": {"Authorization": "Bearer ******"}, "data": "b'" + "0" * 298 + "...",
-                          "method": "GET"})
+        res = mask_opts({"headers": {"Authorization": "Bearer 123456"}, "method": "GET"})
+        self.assertEqual(res, {"headers": {"Authorization": "Bearer ******"}, "method": "GET"})
 
 
 if __name__ == '__main__':